finally commit

+++ src/main/java/postoffice/demo/service/serviceImpl/pressServiceImpl.java
@@ -2,6 +2,7 @@ package postoffice.demo.service.serviceImpl;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import postoffice.demo.dao.NewspaperDao;
 import postoffice.demo.dao.PressDao;
 import postoffice.demo.entity.Press;
 import postoffice.demo.result.ResultMap;
@@ -11,6 +12,8 @@ import postoffice.demo.service.PressService;
 public class pressServiceImpl implements PressService {
     @Autowired
     PressDao pressManager;
+    @Autowired
+    NewspaperDao newspaperManager;
     @Override
     public ResultMap getAll() {
         return ResultMap.data(pressManager.getAll());
@@ -28,7 +31,7 @@ public class pressServiceImpl implements PressService {
                 return ResultMap.errno(-1,"newPress name exist");
             if(pressManager.updateById(press)==1)
                 return ResultMap.errno(0,"success");
-            else return ResultMap.errno(-1,"user does not exist");
+            else return ResultMap.errno(-1,"press does not exist");
         }catch (Exception e){
             return ResultMap.errno(-1,"update error");
         }
@@ -41,7 +44,7 @@ public class pressServiceImpl implements PressService {
                 return ResultMap.errno(-1,"press name exist");
             if(pressManager.insert(press)==1)
                 return ResultMap.errno(0,"success");
-            else return ResultMap.errno(-1,"user does not exist");
+            else return ResultMap.errno(-1,"press does not exist");
         }catch (Exception e)
         {
             return ResultMap.errno(-1,"insert error");
@@ -51,9 +54,11 @@ public class pressServiceImpl implements PressService {
     @Override
     public ResultMap deleteById(int id) {
         try{
+            if(newspaperManager.getByPressId(id)!=null)
+                return ResultMap.errno(-1,"error ,press being used");
             if(pressManager.deleteById(id)==1)
                 return ResultMap.errno(0,"success");
-            else return ResultMap.errno(-1,"user does not exist");
+            else return ResultMap.errno(-1,"press does not exist");
 
         }catch (Exception e){
             return ResultMap.errno(-1,"delete error");
